First try/requirements.txt
pandas
py-AutoClean
pdfkit
dataprep
openpyxl